リファクタリング
トークンを生成するメソッド自体、JWTの機能として準備しておくことが、オブジェクト指向的に正しいので、メソッドをJWTの中に移動。
